+++ backend/src/main/java/net/es/oscars/pss/params/MplsHop.java
@@ -1,0 +1,16 @@
+package net.es.oscars.pss.params;
+
+import lombok.*;
+
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MplsHop {
+
+    private Integer order;
+
+    private String address;
+
+}