+++ backend/src/main/java/net/es/oscars/pss/params/MplsPath.java
@@ -1,0 +1,17 @@
+package net.es.oscars.pss.params;
+
+import lombok.*;
+
+import java.util.List;
+
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class MplsPath {
+
+    private String name;
+
+    private List<MplsHop> hops;
+}