+++ backend/src/main/java/net/es/oscars/pss/params/alu/AluQosType.java
@@ -1,0 +1,6 @@
+package net.es.oscars.pss.params.alu;
+
+public enum AluQosType {
+    SAP_INGRESS,
+    SAP_EGRESS;
+}