+++ shared/src/main/java/net/es/oscars/dto/pss/EthFixtureType.java
@@ -1,7 +1,0 @@
-package net.es.oscars.dto.pss;
-
-public enum EthFixtureType {
-    ALU_SAP,
-    JUNOS_IFCE,
-    REQUESTED
-}