+++ shared/src/main/java/net/es/oscars/dto/pss/cmd/Command.java
@@ -2,10 +2,7 @@ package net.es.oscars.dto.pss.cmd;
 
 
 import lombok.*;
-import net.es.oscars.dto.pss.params.alu.AluParams;
-import net.es.oscars.dto.pss.params.ex.ExParams;
-import net.es.oscars.dto.pss.params.mx.MxParams;
-import net.es.oscars.dto.topo.enums.DeviceModel;
+import net.es.oscars.dto.topo.DeviceModel;
 
 @Data
 @Builder
@@ -24,8 +21,5 @@ public class Command {
     private String connectionId;
     private boolean refresh;
 
-    private AluParams alu;
-    private MxParams mx;
-    private ExParams ex;
 
 }