+++ backend/CHANGES.md
@@ -2,6 +2,7 @@
 
 ### v1.0.35
 > June xx, 2019
+- BREAKING CHANGE: /protected/hold will NOT throw an exception if resources unavailable any more. Callers must check the `validity.valid` field.
 - PSS config generation refactoring 
 - PSS improved queueing, bug fixing