+++ backend/CHANGES.md
@@ -1,5 +1,15 @@
 ## OSCARS Backend Release Notes
 
+### v1.0.35
+> June xx, 2019
+- BREAKING CHANGE: /protected/hold will NOT throw an exception if resources unavailable any more. Callers must check the `validity.valid` field.
+- PSS config generation refactoring 
+- PSS improved queueing, bug fixing
+
+## v1.0.34
+> May 21, 2019
+- Version bump
+
 ### v1.0.33
 > May 15, 2019